use int instead of long to store a reader ID (32 bits)
git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@3271 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/ccid_serial.c
@@ -505,7 +505,7 @@ status_t OpenSerial(unsigned int reader_index, int channel)
 static status_t set_ccid_descriptor(unsigned int reader_index,
 	const char *reader_name, const char *dev_name)
 {
-	long readerID;
+	int readerID;
 	int i;
 	int already_used = FALSE;
 	static int previous_reader_index = -1;
